git commit -m  notification
