+++ package.json
@@ -16,6 +16,7 @@
   "license": "ISC",
   "dependencies": {
     "firebase": "^10.7.1",
+    "izitoast": "^1.4.0",
     "spin.js": "^4.1.1",
     "vite-plugin-full-reload": "^1.0.5",
     "vite-plugin-html-inject": "^1.0.1"
