+++ src/index.html
@@ -5,7 +5,10 @@
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>YOUR ENERGY</title>
-
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/izitoast@1/dist/css/iziToast.min.css"
+    />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/modern-normalize/2.0.0/modern-normalize.min.css"
@@ -45,7 +48,6 @@
 
     <load ="partials/footer.html" />
     <load ="partials/modal.html" />
-
 
     <script type="module" src="/main.js"></script>
   </body>
